Move to a 2.0.0.pre release of the aws-sigv4 gem (TODO: work required to make it exactly conform to existing interface).
