+++ .rubocop.yml
@@ -35,7 +35,7 @@ Naming/FileName:
   Exclude:
     - 'gems/aws-crt/lib/aws-crt.rb'
     - 'gems/aws-crt-auth/lib/aws-crt-auth.rb'
-    - 'gems/aws-crt-signer/lib/aws-crt-signer.rb'
+    - 'gems/aws-sigv4/lib/aws-sigv4.rb'
 
 Style/TrailingCommaInArrayLiteral:
   Enabled: false
