+++ gems/aws-sigv4/aws-sigv4.gemspec
@@ -14,5 +14,6 @@ Gem::Specification.new do |spec|
   spec.platform = Gem::Platform::RUBY
   spec.required_ruby_version = '>= 2.5'
   spec.add_dependency 'aws-crt-auth'
+  spec.add_dependency('aws-eventstream', '~> 1', '>= 1.0.2') # For signing event stream events
   spec.add_development_dependency 'rspec'
 end
