+++ gems/aws-sigv4/lib/aws-sigv4.rb
@@ -1,4 +1,4 @@
 # frozen_string_literal: true
 
 require 'aws-crt-auth'
-require_relative 'aws-crt-signer/signer'
+require_relative 'aws-sigv4/signer'
