+++ gems/aws-sigv4/lib/aws-sigv4/signer.rb
