+++ gems/aws-sigv4/spec/signer_spec.rb
